More Several Changes
Now I changed the Tab1 Main Layout, this time, I had a QGridLayout and I also created some functions to create each
widget iterating through a for loop, where each dictionary has a widget name as a key and its respective position
in the grid as the value. Storing the widgets in another dictionary allows me to access them whenever I want with
loops, so I can set, delete or get their text when I push  the button.

Another big change was the calendar, this is another widget and I open it when the browse button emit the clicked signal
When the calendar is clicked, I retrieve the date in a variable and put this in its Line Widget