Interacting with the database
In this commit, I created the save function that will execute when the user presses the "SAVE" button. This function
iterates through the QLineEdit widgets dictionary and append the value of each widget in a list. Then, I also got the
combobox value but, this time, I store the value in a variable.

Finally, the script gets connected with the database and insert the values in their fields.