Adding the tab6
This new tab inheritss from the database class and the idea is reuse the code so I don't need to create the class
again. In order to reuse the code I set the layout of each tab in the main function and I call the function
that set the table model depending on the SQL Query. Given that each function of the Database class has a different
SQL query, each function is called on its respective instance