Adding the last three tabs
Each function has its own SQL code for query and displays the result in its respective tab.

The code used to get the results of the month also obtains the values in order to put them in a bar graph using
matplotlib. That graph is created with the def grafica(self) function